webChapp/requirements.txt
absl-py==0.8.0
astor==0.8.0
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
image==1.5.27
importlib-metadata==1.5.0
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.2
lxml==4.4.1
Markdown==3.1.1
matplotlib==3.3.1
mccabe==0.6.1
more-itertools==8.2.0
mysql-connector-python==8.0.22
numpy==1.17.2
packaging==20.1
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.9.1
psycopg2==2.8.6
py==1.8.1
pylint==2.4.2
pyparsing==2.4.6
PyQt5==5.13.1
PyQt5-sip==4.19.19
pytest==5.3.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.2
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
tensorboard==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.15.6
whitenoise==5.2.0
wrapt==1.11.2
zipp==2.2.0
